+++ packages/core/src/compileNotAtomicCss.ts
@@ -1,4 +1,5 @@
-import type { StyleOrder, CssRule } from './styleOrder';
+import type { StyleOrder } from './styleOrder';
+import type { CssRule } from './types/common';
 import type { Selectors } from './types/common';
 import type { KeyframesRules, SupportStyle } from './types/style';
 import { compileCss } from './utils/compileCss';
