+++ packages/core/src/index.ts
@@ -7,12 +7,12 @@ export { __style } from './__style';
 export { __globalStyle } from './__globalStyle';
 export { __className } from './__className';
 export { ClassName } from './ClassName';
-export { styleOrder, getStyleOrder, CssRule, StyleOrder } from './styleOrder';
+export { styleOrder, getStyleOrder, StyleOrder } from './styleOrder';
 export { sortCssRules } from './sortCssRules';
 export { uniqueCssRules } from './uniqueCssRules';
 export { setCssRules } from './setCssRules';
 export { isBuildTime } from './isBuildTime';
-export type { Classes, ClassNameOverride, ClassOverride } from './types/common';
+export type { Classes, ClassNameOverride, ClassOverride, CssRule } from './types/common';
 export type { Injector, BuildArgument } from './types/common';
 export type {
   Node,
